+++ components/rpc-daemon/src/commands/who/mod.rs
@@ -164,6 +164,7 @@ async fn request(
             ProtocolResult::NotFound => return Err(who_timeout(pid)),
             ProtocolResult::Unavailable => return Err(unavailable(pid)),
             ProtocolResult::Legend { .. } => return Err(unavailable(pid)),
+            ProtocolResult::Player { .. } => return Err(unavailable(pid)),
         }
     }
 }